+++ src/lib/components/properties-menu/models/fields/property-select-field.model.ts
@@ -4,14 +4,17 @@ import { PropertyOption, PropertyOptionParameters } from '../property-option.mod
 
 export interface PropertySelectFieldParameters<T = unknown> extends Omit<PropertyFieldParameters<T>, 'type'> {
     options?: PropertyOptionParameters<T>[];
+    searchable?: boolean;
 }
 
 export class PropertySelectField<T = unknown> extends PropertyField<T> {
     options: PropertyOption<T>[];
+    searchable: boolean;
 
-    constructor({ options = [], ...base }: PropertySelectFieldParameters<T>) {
+    constructor({ options = [], searchable = false, ...base }: PropertySelectFieldParameters<T>) {
         super({ ...base, type: PropertyFieldType.Select });
 
         this.options = options.map(option => (option instanceof PropertyOption ? option : new PropertyOption(option)));
+        this.searchable = searchable;
     }
 }